requirements.txt
streamlit==1.31.0
ultralytics==8.1.0
opencv-python-headless==4.9.0.80
pillow==10.2.0
numpy==1.26.3
torch==2.5.1
torchvision==0.20.1